[NO GBP] Gives miners general cargo access (#67025)
Gives miners the ability to access their department (like the cargo bay) as normal. This is a quick fix to make sure that miners can do their jobs while we wait for the refactor currently in the works to go through. After that is in I will be able to properly manage the accesses and compartmentalize as necessary.

(cherry picked from commit e5b68960b8df7b8876c559cebf804eb76810da73)