Makes smoke react to opening airlocks. (#66963)
Makes smoke spread through newly opened airlocks if the cloud has spread past the airlock, but has not completely finished spreading.

More intuitive smoke behavior.
As funny as it is to open an airlock and see nothing but a wall of smoke it's even funnier to open an airlock and suddenly be inside a wall of smoke.

(cherry picked from commit 9d16d6643c9fe0aabdc07c8bbbc859885d00c7ba)