[MDB IGNORE] Replaces var access with helpers on away, ruin, and shuttles (#67370)
What it says on the tin.

I edited for req_access with vsc instead of the map editor for our ruins and away missions. It was much more time-efficient! This means I may have accidentally replaced a window access with an access helper (which will not work). I am going to double check that but want a third set of eyes especially on that for whoever reviews this. Thanks.

(cherry picked from commit 5ca39ded065352b3dd8a0f708ed9b3433631ad92)