Organizes the Update Paths tools folder and provides a readme (#67241)
Puts the scripts for Update Paths in a "scripts" subfolder, labels them all according to their associated PR for easy sequential updating on paths, and provides a readme for how to use the tool and how to make the scripts for it.

I don't know what the label would be for this.

* Organizes the Update Paths tools folder and provides a readme

* Clarified adding more than one path update in the readme

* swaps the rtf with a markdown file

* renames the most recent updatepaths and throws it into the folder

(cherry picked from commit 4083eb72e29b9b8b2eca99c0133fbf6894f8c34d)