Fixes duplicated objects due to unresolved merge conflict in IceBox bar (#67698)
Someone reported overlapping objects on a tile on discord so I went and took a look at it. Looks like it was caused by a merge conflict that was not properly resolved (only the marker was deleted). I did my best to delete the objects that were intended to be deleted and resolved the conflict.