Fixes some soulstone issues, overall makes soulstone behavior more consistent (#67846)
Fixes

    Inconsistent behaviour of the cult's sacrifices. #67014

Soulstones had a if(client) check prior to making a shade, which failed if the mob wasn't in their body.
So, we just grab the ghost before making a shade on capture.

This also allowed for the weird hack from #63707 to become unnecessary, so I reverted it.

    Fixes 

    Cult can no longer shade soul-less mobs. #66629

If you used a soulstone on a mindless mob, it would pass in is_sacrifice_target(null), which would always succeed if the cult team had an objective without a target. Which cults do have.

    Overall increases readability and reduces copypaste of soulstone code, making it a tad more consistent.

Moved a lot of copy+pasted "theme" code to the appropriate update_x procs. Cleaned up code in general, as is tradition with cult code.