Fixes admin.fun list (#68336)
About The Pull Request

Many of the items in the list of admin fun were not actually categorized as Admin.Fun (and thus not appearing), and some admin fun procs were not even listed.
Why It's Good For The Game

Basically, more admin fun procs are available that were already in code but not listed
Changelog

cl
admin: Fixed many "admin.fun" listings that were incorrect or missing.
/cl

# Conflicts:
#	code/modules/admin/admin_verbs.dm