[MDB Ignore] Close the Trap Door (#68458)
Replaces one space tile with one plating, where one space and a floor decal was previously. Did not edit the floor decal since it matches the surroundings.

It's an accidental space tile placement under a firelock covered by a sandy floor decal, so it appears to be a floor but in fact is not.