Makes result of decal painter and tile sprayer not cleanable like the map-start decals (#68088)
The decals from decal painters and tile sprayers are identical to the ones from roundstart, except they're washable. This is good for if you need to undo or erase mistakes. Or to janicart clown nonsense quickly. But it sucks if you're trying to cement some proper borders and queue arrows and whatnot in.

So, they now are not washable, but instead you crowbar the floor tiles up, just like with map-start decals.

    Also added ink level meter for toner cartridges.
    Also, replaced my previous half-effort "white" and "black" tile sprayer colors with neutral and dark. Makes neutral and bar burgundy use the correct unusual alpha values as with the real ones.

Why It's Good For The Game

So that these aesthetic improvements aren't accidentally wiped by the janitor cleaning up.

As well, the extremely common neutral tile overlays can now be painted.
Changelog

cl
add: The results of the tile sprayer and decal painter can now only be removed by crowbarring the tile, just like map-start.
fix: Tile sprayer's "white" replaced with "neutral", the most common tile overlay color.
/cl