Allows late-joiners to have a chance at commending other people (#70574)
About The Pull Request

Basically, it was iterating the list of the people that had joined the game, in order. There's only one problem with that: The first entries in that list were always the round-start people, so as soon as you late-join into the shift, you can say goodbye to your chance at commending someone else, which explains why certain players claim to never have had the chance to commend someone else whilst I can oftentimes get the prompt for three or four rounds in a row.

Why It's Good For The Game

Makes the system work as intended, by actually making everyone that joined the round and is still alive at the end of it, a chance to commend others.