515 Compatibility (#71161)
Makes the code compatible with 515.1594+

Few simple changes and one very painful one.
Let's start with the easy:
* puts call behind `LIBCALL` define, so call_ext is properly used in 515
* Adds `NAMEOF_STATIC(_,X)` macro for nameof in static definitions since
src is now invalid there.
* Fixes tgui and devserver. From 515 onward the tmp3333{procid} cache
directory is not appened to base path in browser controls so we don't
check for it in base js and put the dev server dummy window file in
actual directory not the byond root.
* Renames the few things that had /final/ in typepath to ultimate since
final is a new keyword

And the very painful change:
`.proc/whatever` format is no longer valid, so we're replacing it with
new nameof() function. All this wrapped in three new macros.
`PROC_REF(X)`,`TYPE_PROC_REF(TYPE,X)`,`GLOBAL_PROC_REF(X)`. Global is
not actually necessary but if we get nameof that does not allow globals
it would be nice validation.
This is pretty unwieldy but there's no real alternative.
If you notice anything weird in the commits let me know because majority
was done with regex replace.

@tgstation/commit-access Since the .proc/stuff is pretty big change.

Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>
# Conflicts:
#	code/datums/components/crafting/crafting.dm
#	code/datums/components/food/edible.dm
#	code/datums/progressbar.dm
#	code/datums/wounds/bones.dm
#	code/game/atoms_movable.dm
#	code/game/machinery/doors/airlock.dm
#	code/game/objects/structures/crates_lockers/closets.dm
#	code/modules/mob/living/simple_animal/friendly/trader.dm
#	code/modules/projectiles/ammunition/_ammunition.dm
#	code/modules/projectiles/boxes_magazines/_box_magazine.dm
#	code/modules/projectiles/guns/ballistic.dm
#	code/modules/projectiles/guns/energy.dm