Small change to Organic Space Suit purchase text (#71191)
## About The Pull Request

When a changeling activates the organic space suit power, the suit piece
begins dosing the wearer with salbutamol to prevent them from
suffocating in space. You still get the "no oxygen" warning, and there
is no feedback to the player that they're being sustained by chems, so
its a reasonable assumption that some people might not be aware of this.
It's probably a good idea to let the buyer know that their oxygen is
fine beforehand so they don't bother with internals when they aren't
needed.
## Why It's Good For The Game

Helps changelings understand exactly how much bang they'll be receiving
for their bucks.
## Changelog
:cl:
spellcheck: the Organic Space Suit purchase text now informs the buyer
that it regulates oxygen needs automatically.
/:cl: