More logging for Supermatter activations & cleaning?. (#70498)
## About The Pull Request

This pull request adds logging and flags the start of the supermatter
from environmental starts. It also makes the log generation and started
flag flipping a proc. This means that atmospheric starts are now logged
(co2, o2, trit, etc), and so is touching/running into the SM.

[Dreamseeker
9Gkzvyfwdp-1.webm](https://user-images.githubusercontent.com/77420409/195494711-d9295b23-3ec3-447e-b255-fd14454bab5c.webm)


## Why It's Good For The Game

Just a tad more log clarity when you go to investigate. Allows
accidental/intentional/grief starts to be diagnosed easier.