Abductee objective replacement: The Promethean Horror (#71836)
## About The Pull Request

Replaces a kind of really uncomfortable and actively non-interactive
abductee objective with a more interesting one that encourages people to
become a horrifying amalgamation of various bodyparts.

## Why It's Good For The Game

I kept forgetting this was a thing, and boy is is creepy. And not in the
interesting way. This feels more interesting and interactive than hiding
away to apparently give birth and have...nothing happen.

## Changelog
:cl:
add: Replaces a less interesting abductee objective with THE PROMETHEAN
HORROR. The Flesh is Weak, the New Flesh is Not So Weak.
/:cl: