Reworks how legs are rendered yet again because it was very convaluted i hated it (#71721)
## About The Pull Request

The previous method for masking legs properly involved quite heavy icon
manipulation and I didn't like it.
To fix this, I simply made two bodypart layers get used for rendering
legs, a low layer (for the leg that should be hidden behind the other in
a given dir) and the normal bodypart layer.
This still involves icon manipulation, as the unnecessary dirs need to
be cropped properly for the low layer and normal layer, but this should
be a considerably simpler and more reliable system.
Additionally, this fixes the issue of husks not having their legs masked
for whatever reason.

## Why It's Good For The Game

Legs still overlap properly, but the code no longer gives me migraines.

![image](https://user-images.githubusercontent.com/82850673/205526975-8bec2dec-d78f-4409-8e26-56b5a7ea3cc2.png)

![image](https://user-images.githubusercontent.com/82850673/205526992-79356c01-f2f7-4fc3-8753-ce73ca059307.png)

## Changelog

Not player facing, unless something went wrong and i failed to notice.
# Conflicts:
#	code/modules/mob/living/carbon/carbon_update_icons.dm
#	code/modules/surgery/bodyparts/parts.dm