Implements Cooldown On Uploading Books to Newscaster (#72112)
## About The Pull Request

Hey there,

Pretty much on the tin. We implement a cooldown on the backend for a set
amount of time to prevent people spamming the newscaster channel with a
shitload of books (rather than have absolutely no safeguards against
it), and then pipe information to the UI for user feedback.
## Why It's Good For The Game

Fixes #71290

In case you weren't already aware, letting users absolutely wreck the
shit out of player's chat via newscaster spam (all newscasters forcesay
a message when something's uploaded to it) as well as just fill the shit
up of a channel with various vulgarities isn't a really good thing.
## Changelog
:cl:
fix: To prevent spam, there is now a cooldown on being able to upload
assorted books into the newscaster's channel feed.
/:cl:

Let me know if the cooldown seems too long, not hard to change.