Changes a wall on runtimestation to a reinforced one (#72817)
## About The Pull Request

One of the walls in the Runtimestation cargo bay was, amongst all of the
reinforced walls used everywhere else on the station, a regular, normal
wall. I noticed it a few months ago and now compulsively check up on it
every time I open runtime for testing. I think it's time to finally put
it to rest.

I don't know if this was put here for any particular reason. I'm
assuming it's just a goof that went unnoticed because it was covered by
a cargo screen.

![image](https://user-images.githubusercontent.com/28870487/213507536-269ffc97-5fd5-43db-a80b-b2962f387fe5.png)
## Why It's Good For The Game

Consistency in mapping.

Reduces the number of lines in the runtimestation.dmm file by a total of
four. That's an improvement to efficiency on some microscopic,
imperceptible scale, right?
## Changelog
:cl:
fix: a wall covered by a screen on runtimestation has been modified into
a reinforced wall.
/:cl:
# Conflicts:
#	_maps/map_files/debug/runtimestation.dmm