Adds some missing turf_flag defines  (#73007)
## About The Pull Request

Fixes #55151 is fixed it seems and can be closed.

Edit: OK, the carpet bug issue I'm not able to reproduce, I'll chalk it
up to late night brain. but feel free to merge this anyway for the
missing bitfields if you'd like. I'll leave it open.

---

</details>


## Why It's Good For The Game

Adds missing bitfield vars for turf_flags, which allows it to show up in
the vars which leads to less confusion.

## Changelog

:cl:
fix: adds some messing bitfield defines for turf_flags
/:cl: