Shelter for Traveler: Fortified Farming Area (#2257)
Remade farm area to a new one.
Added 3 large buildings:

The largest one is Bar, Gun Shop and some kind of Inn.
"Apartments for poor" is two granaries connected with each other.

Small basement with workshop.

Second farm is "local clinic" with some kind of lab on a second floor.

Third one is a ranch that were rebuilt in a "All-In-One Store". Two floors, most of them uses as warehouse. There also barn for livestock with trails of slaughter in it, and crosses with skeletions of raiders that were killed by Legion.

 And a larger warehouse in a basement.

I also remade Dime's School a little bit.

And added crosses in a front of Legion's Base.

Co-authored-by: Edward Nashton <eddienigma48@gmail.com>