Revert "Underwear Update (Sanity edition)" (#73074)
Reverts tgstation/tgstation#73010

This PR caused problems for underwear code that somehow didn't get
caught in testing. Discovered so far are:

- Spawning at roundstart, latejoin, or by admin tools as a non-human
character removes your underwear #73063
- Taking a head roll with a non-human character results in the human
alternative character also having new underwear
https://github.com/tgstation/tgstation/pull/73010#issuecomment-1408080782

According to @NamelessFairy the root problem here is that on_removal()
gets called during the process of setting up non-human species, leading
to their underwear being removed. The human alternative is presumably a
further extension of this same behaviour.

Given this is likely to require some level of refactoring to avoid, I
think it would be better for everyone to revert until that can be done
and save all non-human players the burden of having to find a dresser
and fix their underwear every time they want to wear it.