Makes arms regrown by the Rod of Asclepius the correct color (#73152)
## About The Pull Request

When an arm holding an activated Rod of Asclepius is cut off, the arm
grows back with the rod held in it. However, due to an oversight, the
arm that grows back is always the "default" color - the albino skintone
for humans/felinids, and grey for everyone else. This PR makes the
regenerated arm have the proper skintone instead.
## Why It's Good For The Game

There's no good reason that a magical doctoring rod should drain all
color from your arm, probably.
## Changelog
:cl:
fix: Corrected the skintone on arms grown back by the Rod of Asclepius
/:cl: