Adds traitor final objective to make the station AI malf (#73020)
## About The Pull Request

Adds a traitor final objective to make the station AI malf. It will give
you a infected law upload that should be used on the AI itself or law
upload console to replace the AI laws with SyndiOS(the person uploading
the laws will be considered as the only syndicate agent), and make it
malfunctional with an objective to protect the uploader(the objective is
cosmetical, since Ai basically is forced to do this by its laws).
Additionaly gives +50 processing time to the malf AI to make things go
faster.
This is tested locally, the objective itself is functionall, as well as
the upload board item.
<details><summary>Screenshots</summary>
<p>


![image](https://user-images.githubusercontent.com/91609255/215290418-8a134a5c-028c-471e-97b4-db24bc0357b5.png)

![image](https://user-images.githubusercontent.com/91609255/215290876-b8c81b11-5fe4-44f2-ac44-7bb87273eb3b.png)

</p>
</details>

Balance-wise this should be okay, since most of the already existing
traitor objectives are more impactful like this(SM cascade, budget
nukies, etc.)

## Why It's Good For The Game

Increases the variety of possible final objectives for traitors to
complete. This one gives fun not only to the traitor themselfes, but
also allows the station silicons to have some fun.

## Changelog
:cl:
add: Added a traitor final objective to infect the station AI with a
virus, that turns it malf and fully loyal to the traitor
/:cl:

---------

Co-authored-by: san7890 <the@san7890.com>