Seed Extractor stores the extracted seeds(On Right Click) in the machine rather than spewing them out on the turf (#73128)
Say you attack the seed extractor with a pumpkin, and it extracts 4
pumpkin seeds. Rather than dumping those 4 seed packets on the turf it
will now store them in the machine.

This feature is optional only if you right click. Left click does the
normal stuff

Saves you the manual work of picking up those packets from the turf and
again putting them in the machine. Keeps the turf clean.