Stabilised Slime Extracts will work from inside MODsuits (#73683)
## About The Pull Request

Fixes #50793 
Fixes #55155 
Fixes #69682

Which actually seemed to work fine except if you were using a mining
MODsuit.
Except that it didn't apply to basic mobs, like bileworms or stick
figures.
I noticed that it was actually pretty hard to tell whether or not the
buff was active in the first place, which might be the real reason for a
bug report... or maybe it literally just didn't work and someone else
fixed it in the past two years. Who can say.

Either way, I did some tidying up here because the order of operations
was a bit messy, and these extracts did not work from inside MODsuits.
As far as I can tell the intended behaviour is that they should be
active from inside hands, pockets, backpacks, boxes in hands, chest
implant, or modsuit storage but not from inside boxes inside backpacks.
So generally, it can be one layer of abstraction away from your body,
but MODsuits were failing on that because technically they're in a suit
which is on your body.

The logic previoulsy added to figure out whether you were in a modsuit
seemed sort of overcomplicated and also didn't work.
I simplified it by just making a snowflake check for modsuits which
overrides the other simplified check to see if you're two levels away
from the mob.
Also the same proc was rewritten in two places which might be why it
didn't work, both places now use the same proc.

I _tried_ to update this to use signals instead of checking every tick,
but the fact that you could put it in a box and then move the _box_ made
it too complicated. It's just less bug-prone to make it check its
location every tick, even if it does make me sad.

Finally I added a balloon alert for when the effect _stops_ applying, so
you can tell why goliaths are suddenly mad at you and not post a bug
report.

## Why It's Good For The Game

Item will now work as expected and give you feedback when it doesn't.
Should be easier to track when effect is and isn't active.

## Changelog

:cl:
fix: Stabilized slime extracts will now once again work from inside
modsuit storage.
fix; Stabilized slime extracts will properly effect carp, spiders,
bileworms, and other "basic mobs".
/:cl: