Fixes the 20+ TC items' discounts being applied incorrectly (#73597)
## About The Pull Request

When #22429 was added, the author's expressed intention was to make 20+
TC items discount less. However, by accident, it has been instead
doubled. I have fixed this problem, while also inverting the values in
`get_discount` to make the calculations easier, and make it clearer what
the return value of the proc is for.

This way, if the 140 TC mauler rolls 75% off, it will be 38% off instead
for 87.5 TC, instead of 87% off for 17 TC.

Also autodocced some procs.

## Why It's Good For The Game

Makes the code work as originally intended.

## Changelog

:cl:
fix: 20+ item's discounts are properly smaller, instead of being larger
/:cl: