Fixes limbs not storing their external organs like they should be (#73951)
## About The Pull Request

That's about it, really. It was messing up the bodytypes, but nobody
noticed because that's not exactly something /tg/ uses much. Downstream,
it's another story, however.

## Why It's Good For The Game
Things going where they're meant to go is definitely a good thing, I
think we can all agree on this.

## Changelog

:cl: GoldenAlpharex
fix: Fixes limbs not properly storing their external organs, resulting
in issues when calculating the bodytypes.
/:cl:
# Conflicts:
#	code/modules/surgery/bodyparts/_bodyparts.dm